refactor development, added exception, uuid field user, customer DTO and modelmapper integration

+++ client-request-service/src/main/java/com/muazwzxv/clients/fraud/FraudClient.java
@@ -8,5 +8,5 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface FraudClient {
 
     @GetMapping("api/v1/fraud-check/{customerId}")
-    FraudCheckResponseDto isFraudster(@PathVariable("customerId") Integer id);
+    FraudCheckResponseDto isFraudster(@PathVariable("customerId") Long id);
 }
